telegram-ai-bot/requirements.txt
aiogram==3.4.1
openai==1.12.0
python-dotenv==1.0.1
httpx==0.26.0
